CI: Add Tianocore edk2 as an external dependency
CI plugins (.pytool/Plugin/) and BaseTools are located in the
Tianocore edk2 repository. To re-use them, add the edk2 repository
as an external dependency.
Once the pip-requirements installed, the stuart_* set of commands
will download the required dependencies to locally run the CI. The
edk2 repository and its dependencies will be downloaded.

It is not currently possible to specify a branch name in the "version"
field. Thus, a fixed commit hash (current edk2 master) is written.
It will eventually be replaced by 'master'.

Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Leif Lindholm <leif@nuviainc.com>
Cc: Sean Brogan <sean.brogan@microsoft.com>
Cc: Bret Barkelew <Bret.Barkelew@microsoft.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Cc: Sami Mujawar <sami.mujawar@arm.com>
Signed-off-by: Pierre Gondois <Pierre.Gondois@arm.com>